Move types to a separated file
